change recommedation algorithm with weight, add crud operation for user,answer,question,... finish auth with jwt feature

+++ src/main/java/net/duchung/quora/service/impl/AnswerServiceImpl.java
@@ -6,6 +6,7 @@ import net.duchung.quora.dto.response.UserResponseDto;
 import net.duchung.quora.entity.Answer;
 import net.duchung.quora.entity.Question;
 import net.duchung.quora.entity.User;
+import net.duchung.quora.exception.AccessDeniedException;
 import net.duchung.quora.exception.DataNotFoundException;
 import net.duchung.quora.mapper.AnswerMapper;
 import net.duchung.quora.mapper.BaseMapper;
@@ -13,12 +14,14 @@ import net.duchung.quora.repository.AnswerRepository;
 import net.duchung.quora.repository.QuestionRepository;
 import net.duchung.quora.repository.UserRepository;
 import net.duchung.quora.service.AnswerService;
+import net.duchung.quora.service.AuthService;
 import org.mapstruct.factory.Mappers;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -30,21 +33,16 @@ public class AnswerServiceImpl implements AnswerService {
     QuestionRepository questionRepository;;
     @Autowired
     UserRepository userRepository;
+    @Autowired
+    AuthService authService;
     @Override
     @Transactional
     public AnswerDto createAnswer(AnswerDto answerDto) {
-        Optional<User> userOpt = userRepository.findById(answerDto.getUserId());
-        Optional<Question> questionOpt = questionRepository.findById(answerDto.getQuestionId());
-        if(userOpt.isEmpty()) {
-            throw new DataNotFoundException("User with id "+answerDto.getUserId()+" not found");
-        }
 
-        if(questionOpt.isEmpty()) {
-            throw new DataNotFoundException("Question with id "+answerDto.getQuestionId()+" not found");
-        }
+        User user = authService.getCurrentUser();
+        Question question = questionRepository.findById(answerDto.getQuestionId()).orElseThrow(() -> new DataNotFoundException("Question with id "+answerDto.getQuestionId()+" not found"));
 
-        Answer answer = new Answer(answerDto.getContent(), 0,questionOpt.get(), userOpt.get(), Collections.emptySet(),Collections.emptySet());
-
+        Answer answer = new Answer(answerDto.getContent(), 0,question, user, Collections.emptySet(),Collections.emptySet());
         Answer savedAnswer = answerRepository.save(answer);
         return toDto(savedAnswer);
     }
@@ -54,31 +52,28 @@ public class AnswerServiceImpl implements AnswerService {
 
     public AnswerDto updateAnswer(Long id, AnswerDto answerDto) {
 
-        if(!questionRepository.existsById(answerDto.getQuestionId())) {
-            throw new DataNotFoundException("Question with id "+answerDto.getQuestionId()+" not found");
-        }
-        if(!userRepository.existsById(answerDto.getUserId())) {
-            throw new DataNotFoundException("User with id "+answerDto.getUserId()+" not found");
-        }
-        if(!answerRepository.existsById(id)) {
-            throw new DataNotFoundException("Answer with id "+id+" not found");
+        User user = authService.getCurrentUser();
+        Answer answer = answerRepository.findById(id).orElseThrow(() -> new DataNotFoundException("Answer with id "+id+" not found"));
 
-        }else {
-            Answer answer = answerRepository.findById(id).get();
-
-            answer.setContent(answerDto.getContent());
-            answer.setQuestion(questionRepository.findById(answerDto.getQuestionId()).get());
-            answer.setUser(userRepository.findById(answerDto.getUserId()).get());
-            Answer savedAnswer = answerRepository.save(answer);
-
-            return toDto(savedAnswer);
+        if(!answer.getUser().getId().equals(user.getId())) {
+            throw new AccessDeniedException("You don't have permission to update this answer");
         }
+        answer.setContent(answerDto.getContent());
+
+        Answer savedAnswer = answerRepository.save(answer);
+
+        return toDto(savedAnswer);
+
     }
 
     @Override
     @Transactional
-
     public void deleteAnswerById(Long id) {
+        Long userId = authService.getCurrentUser().getId();
+        Answer answer = answerRepository.findById(id).orElseThrow(() -> new DataNotFoundException("Answer with id "+id+" not found"));
+        if(!answer.getUser().getId().equals(userId)) {
+            throw new AccessDeniedException("You don't have permission to delete this answer");
+        }
         answerRepository.deleteById(id);
     }
 
@@ -93,8 +88,9 @@ public class AnswerServiceImpl implements AnswerService {
         BaseMapper.getBaseDtoAttribute(answerDto,answer);
         return answerDto;
     }
-    private Answer toEntity(AnswerDto answerDto) {
-        Answer answer = ANSWER_MAPPER.toAnswer(answerDto);
-        return answer;
+
+    @Override
+    public List<AnswerDto> getAnswersByQuestionId(Long questionId) {
+        return answerRepository.findByQuestionId(questionId).stream().map(this::toDto).toList();
     }
 }
